support for CPLD (really FPGA) register values that are not as wide as the value that is specified in hardware. this stuff is for backward compatibility in part and also because the value sent is addressed to a particular register and the value is sent as at least a byte value or multiple of byte number of bits. So it is desireable to specify as well the max actual value that the hardware can support.  backfitted without careful checking to existing classes. Made some of the ApsReadoutControl registers 6 bit values but didn't check them yet.
git-svn-id: https://svn.code.sf.net/p/jaer/code/jAER/trunk@6225 b7f4320f-462c-0410-a916-d9f35bb82d52

+++ src/eu/seebetter/ini/chips/davis/DavisBaseCamera.java
@@ -453,8 +453,8 @@ abstract public class DavisBaseCamera extends DavisChip implements RemoteControl
                 if (newExposure == currentExposure) {
                     newExposure++; // ensure increase
                 }
-                if (newExposure > getDavisConfig().getExposureControlRegister().getMax()) {
-                    newExposure = getDavisConfig().getExposureControlRegister().getMax();
+                if (newExposure > (int)getDavisConfig().getExposureControlRegister().getMax()) {
+                    newExposure = (int)getDavisConfig().getExposureControlRegister().getMax();
                 }
                 if (newExposure != currentExposure) {
                     getDavisConfig().getExposureControlRegister().set(newExposure);
@@ -466,7 +466,7 @@ abstract public class DavisBaseCamera extends DavisChip implements RemoteControl
                     newExposure--; // ensure decrease even with rounding.
                 }
                 if (newExposure < getDavisConfig().getExposureControlRegister().getMin()) {
-                    newExposure = getDavisConfig().getExposureControlRegister().getMin();
+                    newExposure = (int)getDavisConfig().getExposureControlRegister().getMin();
                 }
                 if (newExposure != currentExposure) {
                     getDavisConfig().getExposureControlRegister().set(newExposure);
